fix expecatation validation
