+++ getting_started_tutorial_final_v3_api/run_checkpoint.py
@@ -4,7 +4,7 @@ import great_expectations as ge
 import sys, os
 sys.path.append(os.path.join(os.path.dirname(os.path.realpath(__file__)), 'great_expectations'))
 
-from plugins import column_custom_max_expectation, column_custom_max_metric, column_custom_max_config_validation
+from plugins import column_custom_max_expectation, column_custom_max_metric
 
 context = ge.get_context()
 context.run_checkpoint(checkpoint_name="my_checkpoint_with_custom_expectation")
